unsigned int -> int64_t

+++ src/libguac/guacamole/protocol.h
@@ -38,6 +38,7 @@
 
 #include <cairo/cairo.h>
 #include <stdarg.h>
+#include <stdint.h>
 
 /* CONTROL INSTRUCTIONS */
 
@@ -623,7 +624,7 @@ int guac_protocol_send_arc(guac_socket* socket, const guac_layer* layer,
 /**
  * Send an audit msg for the activity log
  */
-int guac_protocol_audit_msg(guac_socket* socket, char *msg, unsigned int len);
+int guac_protocol_audit_msg(guac_socket* socket, char *msg, int64_t len);
 
 
 /**
